+++ komikaan.Irrigator/Models/PsqlStopTimeUpdate.cs
@@ -3,11 +3,11 @@
     public class PsqlStopTimeUpdate
     {
         public string DataOrigin { get; set; }
-        public Guid InternalId { get; set; } = Guid.NewGuid();
+        public string Id { get; set; }
         public DateTimeOffset LastUpdated { get; set; } = DateTimeOffset.UtcNow;
-        public int StopSequence { get; set; }
-        public string StopId { get; set; }
-        public required string TripId { get; set; }
+        public int? StopSequence { get; set; }
+        public string? StopId { get; set; }
+        public string? TripId { get; set; }
         public int? ArrivalDelay { get; set; }
         public DateTimeOffset? ArrivalTime { get; set; }
         public int? ArrivalUncertainty { get; set; }